refactor(piw-utils-internal): refactor getDimensions util function into piw-utils-internal
