+++ packages/tools/piw-utils-internal/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./dom";
 export * from "./PageEditorUtils";
 export * from "./DataGridFilterIcons";
+export * from "./StylingPropertyUtils";
